pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/src/takehome"]

[project]
name = "orbital-takehome"
version = "0.1.0"
description = "Document Q&A tool for commercial real estate lawyers"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    # LLM
    "pydantic-ai-slim[anthropic]>=0.0.39",
    # Document extraction
    "pymupdf>=1.25.0",
    # HTTP client
    "httpx>=0.27.0",
    # Settings
    "pydantic-settings>=2.0.0",
    # Logging
    "structlog>=25.1.0",
    # File upload
    "python-multipart>=0.0.9",
    # Environment
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["takehome"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
extraPaths = ["backend/src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
